requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dask==2.19.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.8.3
h5py==2.10.0
hyperspy==1.5.2
idna==2.9
imageio==2.8.0
ipykernel==5.3.0
ipyparallel==6.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
llvmlite==0.33.0
MarkupSafe==1.1.1
matplotlib==3.2.2
matplotlib-scalebar==0.6.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
mpmath==1.1.0
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numba==0.50.0
numexpr==2.7.1
numpy==1.19.0
opencv-python==4.2.0.34
openpyxl==3.0.3
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
Pint==0.13
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
PTable==0.9.2
ptyprocess==0.6.0
py==1.8.2
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.16.0
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
scikit-image==0.17.2
scipy==1.5.0
Send2Trash==1.5.0
six==1.15.0
sparse==0.10.0
statsmodels==0.11.1
sympy==1.6
temmeta==0.0.4
terminado==0.8.3
testpath==0.4.4
tifffile==2020.6.3
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
traits==6.1.0
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
